move scripts into scripts folder
